# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alina group test project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/task/get": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Task list",
                "operationId": "task.list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "requestBody": {
                    "description": "Filter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "title #1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description #1"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "low"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "active"
                                    },
                                    "expired_at": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2024-04-20 17:42:16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Unauthenticated"
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": "..."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/task/getById/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Task getById",
                "operationId": "task.getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Unauthenticated"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Task not found"
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "id": 1,
                                            "title": "title #1",
                                            "description": "description #1",
                                            "priority": "low",
                                            "status": "active",
                                            "expired_at": "2024-04-20 17:42:16",
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/task/create": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Task create",
                "operationId": "task.create",
                "requestBody": {
                    "description": "Create parameters [prority: low, meduim, high]",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "priority",
                                    "status",
                                    "expired_at"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "title #1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description #1"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "high"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "active"
                                    },
                                    "expired_at": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2024-04-20 17:42:16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "id": 100,
                                            "title": "title #1",
                                            "description": "description #1",
                                            "priority": "high",
                                            "status": "active",
                                            "expired_at": "2024-04-20 17:42:16",
                                            "created_at": "2024-04-21T08:23:55.000000Z",
                                            "updated_at": "2024-04-21T08:23:55.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Unauthenticated"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/task/update/{id}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Task update",
                "operationId": "task.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update task data [prority: low, meduim, high]",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "title #1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description #1"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "low"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "active"
                                    },
                                    "expired_at": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2024-04-20 17:42:16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "id": 100,
                                            "title": "title #1",
                                            "description": "description #1",
                                            "priority": "high",
                                            "status": "active",
                                            "expired_at": "2024-04-20 17:42:16",
                                            "created_at": "2024-04-21T08:23:55.000000Z",
                                            "updated_at": "2024-04-21T08:23:55.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Task not found"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Unauthenticated"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/task/delete/{id}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Task delete",
                "operationId": "task.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Task deleted"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Unauthenticated"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}